feat(#757): add activity objective formatting util
